reddit_trends_analytics/requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
confluent-kafka==2.3.0
delta-spark==2.1.0
fastavro==1.9.0
idna==3.6
importlib-metadata==7.0.0
minio==7.2.0
py4j==0.10.9.5
pycparser==2.21
pycryptodome==3.19.0
pyspark==3.3.1
python-dotenv==1.0.0
requests==2.31.0
urllib3==2.1.0
zipp==3.17.0
